requirements.txt
# SUE Improved Dependencies

# Core
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
umap-learn>=0.5.0

# Progress bars
tqdm>=4.62.0

# Optional: for faster nearest neighbors
# annoy>=1.17.0